refactor: core classes restructure
