+++ RSSFeedify/RSSFeedifyClientCore/Services/Networking/IAuthenticationHeader.cs
@@ -1,4 +1,4 @@
-﻿namespace RSSFeedifyClientCore
+﻿namespace RSSFeedifyClientCore.Services.Networking
 {
     public interface IAuthenticationHeader
     {
